# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aftersales Api - SP21",
        "description": "List Aftersales Api",
        "version": "1.0"
    },
    "paths": {
...
        "/insurance": {
            "post": {
                "tags": [
                    "Insurance"
                ],
                "summary": "Create the insurance for a product",
                "operationId": "e887c6f6daad2502cddb58f520df1280",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "content",
                                    "start_date",
                                    "expired_date"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "number",
                                        "example": "4"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Trong tuần đầu tiên, bạn có thể đổi một sản phẩm bất kỳ có giá tiền bằng hoặc hơn sản phẩm đã mua "
                                    },
                                    "start_date": {
                                        "type": "date",
                                        "example": "2022-01-01"
                                    },
                                    "expired_date": {
                                        "type": "date",
                                        "example": "2022-01-08"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Success": {
                                            "type": "string",
                                            "example": "Create the insurance for product successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "API key is missing."
                    }
                }
            }
        },
        "/product/{id}/insurance": {
            "get": {
                "tags": [
                    "Insurance"
                ],
                "summary": "Get the insurance of product has id = $id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "insurances": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "product_id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "content": {
                                                        "type": "number",
                                                        "example": "Bảo hành một đổi một trong vòng 1 tháng đầu sử dụng"
                                                    },
                                                    "start_date": {
                                                        "type": "date",
                                                        "example": "2022-01-01"
                                                    },
                                                    "expired_date": {
                                                        "type": "date",
                                                        "example": "2022-02-01"
                                                    },
                                                    "created_at": {
                                                        "type": "timestamp",
                                                        "example": "2022-01-20T09:06:33.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "timestamp",
                                                        "example": "2022-01-20T09:06:33.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "API key is missing."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Insurance"
                ],
                "summary": "Delete the insurance",
                "description": "Delete the insurance by product_id",
                "operationId": "0fadde1271ad73a6801a2858b1235ef6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete the insurance successfully"
                    },
                    "403": {
                        "description": "API key is missing."
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}